Rename function folder
- rename folder to match other naming convention
- change import path in App.tsx
